Add info on functionality, endpoints, versioning
Updated the readme with some info about the generator and other
endpoints, because this info probably will be useful both for the
maintainers of the repo and possibly other users as well.

Also added a tiny bit of info about the versioning strategy which was
selected in this iteration. The versioning strategy will be up for
discussion, but this is a first version and deserves a mention in the
README. :)