Refactor tests
Why:
The existing test was a bit cluttered with input and output values. To
simplify writing new tests, we wanted to move those to separate files.

How: We created separate files containing valid inputs and outputs, and
used the created methods to include in the test.

Co-authored-by: ssbdif <35797607+ssbdif@users.noreply.github.com>
Co-authored-by: larsroger <lrb@ssb.no>

int.ref: [STRATUS-621]
int.ref: [STRATUS-802]