Introduce health and metrics endpoint definitions
Adding definitions for K8s health probes and metrics.

We introduced parameters so people can choose to include k8s probe
definitions and/or metrics endpoint definitions in the HelmRelease.
We did this because the values used (paths) are hardcoded for now so
people can't specify the path their app actually uses (if any), and if
the health probes fail due to the endpoints not existing or being wrong,
K8s will not be able to orchestrate the app properly. (It will probably
try to restart the pod a gazillion times because the app never responds
that it is alive and/or healthy.)

Internal ref [STRATUS-745]

Co-authored-by: ArielKarlsen <54323769+ArielKarlsen@users.noreply.github.com>