Add input validation
Why:
We want to ensure that the service creates a valid HelmRelease.

How:
We've added constraints on the input fields in the model for the
HelmRelease. We've tested that the framework (Pydantic) returns
meaningful error messages if validation fails.

Tests were added to ensure that the framework behaves as we
expected, but not all possible erroneous inputs are under
test. Please add tests as suited when adding new
fields to the model.

Some metadata in the docs page were added to give more
information about the application.

An entrypoint which return a json-schema was created
to allow for easy usage of the json-schema in clients.

Internal ref: [STRATUS-804]

Co-authored-by: ssbdif <35797607+ssbdif@users.noreply.github.com>
Co-authored-by: are3k <5612258+are3k@users.noreply.github.com>